requirements.txt
# Core Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0

# WebSocket support
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.34.2

# Google Gemini AI
google-generativeai>=0.8.0

# Background tasks
APScheduler==3.10.4

# Environment management
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
